requirements.txt
google-genai==1.57.0
neo4j==5.28.2
ollama==0.6.0
pandas==2.3.3
pydantic==2.12.2
python-dotenv==1.2.1
selenium==4.39.0
SPARQLWrapper==2.0.0
tqdm==4.67.1
fastapi==0.128.0
uvicorn==0.40.0
